--- multiversx_sdk_core/transaction_factories/__init__.py
@@ -1,15 +1,0 @@
-from multiversx_sdk_core.transaction_factories.delegation_factory import \
-    DelegationTransactionIntentsFactory
-from multiversx_sdk_core.transaction_factories.smart_contract_factory import \
-    SmartContractTransactionIntentsFactory
-from multiversx_sdk_core.transaction_factories.token_operations_factory import \
-    TokenOperationsTransactionIntentsFactory
-from multiversx_sdk_core.transaction_factories.transaction_factory_config import \
-    TransactionFactoryConfig
-
-__all__ = [
-    "DelegationTransactionIntentsFactory",
-    "TokenOperationsTransactionIntentsFactory",
-    "TransactionFactoryConfig",
-    "SmartContractTransactionIntentsFactory"
-]
