+++ multiversx_sdk_core/transaction_intents_factories/smart_contract_transaction_intents_factory.py
@@ -7,7 +7,7 @@ from multiversx_sdk_core.constants import (CONTRACT_DEPLOY_ADDRESS,
                                            VM_TYPE_WASM_VM)
 from multiversx_sdk_core.interfaces import IAddress
 from multiversx_sdk_core.serializer import arg_to_string, args_to_strings
-from multiversx_sdk_core.transaction_factories.transaction_intent_builder import \
+from multiversx_sdk_core.transaction_intents_factories.transaction_intent_builder import \
     TransactionIntentBuilder
 from multiversx_sdk_core.transaction_intent import TransactionIntent
